refactors table view to account for new data shape
