+++ src/components/table/TableView.svelte
@@ -28,8 +28,7 @@ export let bucketTypeLabel = 'Categories';
 
 let totalPages = 0;
 let currentPage = 0;
-let selectedData = data[currentKey][currentAggregation];
-$: selectedData = data[currentKey][currentAggregation];
+let selectedData = data;
 $: if (selectedData) currentPage = 0;
 $: totalPages = Math.ceil(selectedData.length / pageSize);
 
@@ -40,8 +39,8 @@ $: largestAudience = Math.max(...selectedData.map((d) => d.audienceSize));
 
 <style>
 span.h {
-  font-weight: bold; 
-  color: var(--cool-gray-700); 
+  font-weight: bold;
+  color: var(--cool-gray-700);
   font-size: var(--text-01);
 }
 
@@ -52,7 +51,7 @@ span.bucket {
 
 <div style="border-bottom: var(--space-base) solid var(--cool-gray-100);">
   <div style="
-    margin-top: var(--space-2x); 
+    margin-top: var(--space-2x);
     margin-bottom: var(--space-2x);
     padding-left: var(--space-4x);
     padding-right: var(--space-4x);
@@ -63,7 +62,7 @@ span.bucket {
   </div>
   <DataTable overflowX={true}>
     <thead>
-      
+
       <Row header>
         <Cell colspan={2} freezeX bottomBorder={false}></Cell>
         <Cell colspan={2} align=left freezeX bottomBorder={false}>
@@ -74,7 +73,7 @@ span.bucket {
       </Row>
 
       <Row header>
-        <Cell 
+        <Cell
           backgroundColor=var(--cool-gray-subtle)
           topBorder={true}
           bottomBorderThickness=2px freezeX size=max tooltip="the {aggregationLevel === 'build_id' ? ' build id' : 'version' } associated with this row">
@@ -92,8 +91,8 @@ span.bucket {
           rightBorder
           backgroundColor=var(--cool-gray-subtle)
           bottomBorderThickness=2px
-          freezeX 
-          align=left 
+          freezeX
+          align=left
           tooltip="the total number of clients associated with this {aggregationLevel === 'build_id' ? ' build id' : 'version' }">
           <span class=h>
             Clients
