+++ src/utils/probe-utils.js
@@ -174,6 +174,20 @@ export function clientCounts(arr) {
   return arr.map((a) => ({ totalClients: a.audienceSize, label: a.label }));
 }
 
+function uniques(d, k) {
+  return Array.from(new Set(d.map((di) => di[k])));
+}
+
+export function gatherProbeKeys(d) {
+  // return Object.keys(nestedData);
+  return uniques(d, 'metric_key');
+}
+
+export function gatherAggregationTypes(d) {
+  // return Object.keys(Object.values(nestedData)[0]);
+  return uniques(d, 'client_agg_type');
+}
+
 export function isSelectedProcessValid(processes, selectedProcess) {
   let process = selectedProcess;
   if (process === 'all_childs') {
